Create `Yiisoft\Yii\Cycle\Schema` namespace; Conveyor classes and Generator moved there
