+++ tests/Schema/Conveyor/AnnotatedSchemaConveyorTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Yiisoft\Yii\Cycle\Tests\Conveyor;
+namespace Yiisoft\Yii\Cycle\Tests\Schema\Conveyor;
 
 use PHPUnit\Framework\TestCase;
-use Yiisoft\Yii\Cycle\Conveyor\AnnotatedSchemaConveyor;
 use Yiisoft\Yii\Cycle\Exception\EmptyEntityPathsException;
-use Yiisoft\Yii\Cycle\Tests\Conveyor\Stub\FakeContainer;
-use Yiisoft\Yii\Cycle\Tests\Conveyor\Stub\FakeGenerator;
+use Yiisoft\Yii\Cycle\Schema\Conveyor\AnnotatedSchemaConveyor;
+use Yiisoft\Yii\Cycle\Tests\Schema\Conveyor\Stub\FakeContainer;
+use Yiisoft\Yii\Cycle\Tests\Schema\Conveyor\Stub\FakeGenerator;
 
 class AnnotatedSchemaConveyorTest extends TestCase
 {
