+++ tests/Schema/Conveyor/SchemaConveyorTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Yiisoft\Yii\Cycle\Tests\Conveyor;
+namespace Yiisoft\Yii\Cycle\Tests\Schema\Conveyor;
 
 use Cycle\Schema\GeneratorInterface;
 use PHPUnit\Framework\TestCase;
-use Yiisoft\Yii\Cycle\Conveyor\SchemaConveyor;
 use Yiisoft\Yii\Cycle\Exception\BadGeneratorDeclarationException;
-use Yiisoft\Yii\Cycle\Tests\Conveyor\Stub\FakeContainer;
-use Yiisoft\Yii\Cycle\Tests\Conveyor\Stub\FakeGenerator;
+use Yiisoft\Yii\Cycle\Schema\Conveyor\SchemaConveyor;
+use Yiisoft\Yii\Cycle\Tests\Schema\Conveyor\Stub\FakeContainer;
+use Yiisoft\Yii\Cycle\Tests\Schema\Conveyor\Stub\FakeGenerator;
 
 class SchemaConveyorTest extends TestCase
 {
